requirements.txt
spotipy
boto3
jupyter
pydantic
python-dotenv